--- PickFLPrj/src/com/kh/inquery/controller/InqueryController.java
@@ -1,18 +1,0 @@
-package com.kh.inquery.controller;
-
-import java.io.IOException;
-
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-@WebServlet("/inquery")
-public class InqueryController extends HttpServlet{
-	
-	@Override
-	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-		req.getRequestDispatcher("/WEB-INF/views/inquery/inquery.jsp").forward(req, resp);
-	}
-}
