+++ PickFLPrj/src/com/pickfl/admin/controller/AdminMemberModify.java
@@ -1,4 +1,4 @@
-package com.kh.admin.controller;
+package com.pickfl.admin.controller;
 
 import java.io.IOException;
