+++ PickFLPrj/src/com/pickfl/member/model/dao/MemberDao.java
@@ -1,6 +1,7 @@
 package com.pickfl.member.model.dao;
 
 import java.sql.Connection;
+
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -8,7 +9,7 @@ import java.sql.Timestamp;
 import java.util.ArrayList;
 import java.util.List;
 
-import static com.pickfl.common.JDBCTemplate.*;
+import static com.pickfl.common.JdbcTemplate.*;
 import com.pickfl.member.model.vo.MemberVo;
 
 public class MemberDao {
